PTTM.html
This is my very first website. It's for nuzzlockes (or however it's written) pokemon playing style people. It basically gives you a list of types. This list represents the only types you can have in your team. The list is either made so that you are able to be supereffective agasint all types of pokemon or the opposite (your team can be resistant to all of the types). That's it. There's not much to it. Hope somebody enjoys it.